+++ light-type-random-gradation.h
@@ -12,8 +12,8 @@ class LightTypeRandomGradation : public LightType
     LightTypeRandomGradation(Adafruit_NeoPixel* strip, uint16_t speedMs);
     virtual void init();
     virtual void loop();
-    virtual void setParam(int speedMs) {_speedMs = speedMs;};
-    virtual int getParam() {return _speedMs;};
+    virtual void setParam(uint16_t speedMs) {_speedMs = speedMs;};
+    virtual uint16_t getParam() {return _speedMs;};
     
   private:
     double diffAbs(byte last, byte current);
